Where do I start lol
So its like 6:45 AM and it's finally done to satisfaction. So first off,
I just changed the name of the file so it doesn't mess with what Sy
added. You would have to change it back to Calendar.java to run it. I
know there's a branching thing in Git for it but I'm not gonna bother
learning that at this hour. So I built most of the functionality into a
separate class so it would allow for testing and modularity and
whatever. There's a bunch of arrays in this class that hold the data.
You can look through it and probably see whats going on with it. I made
a few small changes to the main calendar file. I rearranged the original
buttons a little bit and made it so that people can choose the name of
the file they created. How I placed my event implementation is to have a
button pressed to add a file to the EventReader class (class which is
used to sort events and calculate great circle distance). After all the
events are added then another button is pressed to calculate the
distance. This prints to the system.out and also writes the comment to
the files that were added. I've put in various tests for correct input
and input that matches what we're looking for but I haven't done a whole
lot of testing (kinda running out of time as I have to be at work in a
hour). So what I have it doing now is the user can keep writing new
event files, and then reading them to the event reader. Anyway yeah just
check it out.