a few minor changes
I noticed when the the program would read in a file, it wouldn't write
out to the same file it read. It create a new file which was the
SUMMARY+.ics. I modified the program so that it writes to the same file
it reads in by basically adding a additional variable into the IcsEvent
object that held the original file name.